refactor: rename pkg githubrelease to githubreleases
